Split BootReceiver, and make both halves survive onReceive returning
`BootReceiver` was 812 lines doing two unrelated jobs: boot recovery, and
delivering five kinds of alarm. Alarm delivery is now `PrayerAlarmReceiver`;
what is left is ~110 lines of recovery.

The split is load-bearing for what comes next. `BootReceiver` calls
`WidgetUpdateScheduler.ensureScheduled`, and that one line pins it to `:app`:
a `:core:*` module naming it is a `:core:*` -> `:feature:*` edge and
`moduleBoundary` fails the build on it. Nothing in `PrayerAlarmReceiver` names
anything in `:feature:widget`, so the notification stack can move.

**goAsync().** Neither half had it, and every branch of both starts work that
outlives `onReceive` — after which Android is free to decide the process has
nothing to do and kill it, most likely right after a reboot, which is exactly
when nothing else will do that work. The receiver-field `CoroutineScope` was
never cancelled either, because the instance is discarded after each call; a
scope per dispatch, cancelled in `finally` beside `pendingResult.finish()`, is
both correct and shorter. `goAsync()` returns null when the receiver was not
dispatched by the framework, so the result is nullable rather than asserted —
which is also what keeps the existing unit tests driving `onReceive` directly.

**PendingIntent safety.** An alarm holds a `PendingIntent`, and a `PendingIntent`
names a component: every alarm armed by an older build names `BootReceiver` and
would now fire into a receiver that no longer handles its action, silently.
`AppInitializer` re-arms on launch, so anyone who opens the app self-heals;
`MY_PACKAGE_REPLACED` — which the repo had no filter for anywhere — closes the
rest, because the OS takes the app out of the stopped state to deliver it and
the re-arm happens seconds after the update whether or not the app is opened.

**Three dead branches.** `LOCKED_BOOT_COMPLETED` and both `QUICKBOOT_POWERON`
spellings were handled in code with no manifest filter behind them, so they
never arrived once. The QUICKBOOT actions are declared now — they are what some
OEM fast-boot implementations send instead of `BOOT_COMPLETED`, and the work
they trigger needs nothing special. `LOCKED_BOOT_COMPLETED` is removed instead:
making it arrive needs `directBootAware="true"`, and the reschedule reaches
DataStore and Room, both unreadable before first unlock.

`BootReceiverManifestTest` is new and is why that will not recur. A receiver's
contract is half Kotlin and half XML and nothing in the build checked that the
two halves agree; it checks both directions, and the old state fails it.

The five alarm filters `BootReceiver` declared are gone with the split — those
actions have been delivered by explicit `PendingIntent` for a long time, so the
filters only meant an implicit broadcast of one could reach the receiver.
