+++ app/src/androidTest/java/com/arshadshah/nimaz/support/HiltTestRunner.kt
@@ -6,6 +6,7 @@ import android.content.Context
 import android.content.pm.PackageManager
 import androidx.test.runner.AndroidJUnitRunner
 import com.arshadshah.nimaz.core.util.BootReceiver
+import com.arshadshah.nimaz.core.util.PrayerAlarmReceiver
 import dagger.hilt.android.testing.HiltTestApplication
 
 /**
@@ -28,9 +29,9 @@ class HiltTestRunner : AndroidJUnitRunner() {
         className: String?,
         context: Context?,
     ): Application {
-        // Disable the manifest-registered BootReceiver for the test process.
+        // Disable the two manifest-registered receivers for the test process.
         //
-        // BootReceiver is an @AndroidEntryPoint whose generated onReceive() injects
+        // Each is an @AndroidEntryPoint whose generated onReceive() injects
         // Hilt fields *before* our code runs. Under HiltTestApplication the Hilt
         // component doesn't exist until a test's HiltAndroidRule.inject() creates it,
         // so any broadcast delivered before the first test — e.g. a past-due exact
@@ -38,18 +39,20 @@ class HiltTestRunner : AndroidJUnitRunner() {
         // prayer notifications — injects into a missing component and crashes the whole
         // instrumentation process ("The component was not created"), running 0 tests.
         //
-        // No test exercises BootReceiver (boot/alarm rescheduling isn't under test;
-        // scheduler logic is verified directly), so disabling it here — on the main
+        // No instrumented test exercises either (boot recovery and alarm delivery aren't
+        // under test; scheduler logic is verified directly), so disabling them here — on the main
         // thread during bindApplication, before the looper dispatches any queued
         // broadcast — removes the crash deterministically without affecting the
         // shipped app. Best-effort: never let this fail the run.
         context?.let { ctx ->
-            runCatching {
-                ctx.packageManager.setComponentEnabledSetting(
-                    ComponentName(ctx, BootReceiver::class.java),
-                    PackageManager.COMPONENT_ENABLED_STATE_DISABLED,
-                    PackageManager.DONT_KILL_APP,
-                )
+            listOf(BootReceiver::class.java, PrayerAlarmReceiver::class.java).forEach { receiver ->
+                runCatching {
+                    ctx.packageManager.setComponentEnabledSetting(
+                        ComponentName(ctx, receiver),
+                        PackageManager.COMPONENT_ENABLED_STATE_DISABLED,
+                        PackageManager.DONT_KILL_APP,
+                    )
+                }
             }
         }
         return super.newApplication(cl, HiltTestApplication::class.java.name, context)
