+++ docs/ARCHITECTURE.md
@@ -215,6 +215,18 @@ below `:core:data` can want it.) `BootReceiver`,
 `PrayerRescheduler`, `InAppUpdateManager` and `core/init` stay in `:app` permanently — a manifest
 entry point and a composition root are app concerns.
 
+**`BootReceiver` is now only the first of those.** It was 812 lines doing two unrelated jobs: boot
+recovery, and delivering five kinds of alarm. Alarm delivery is `core/util/PrayerAlarmReceiver.kt`
+now; what is left is ~110 lines of recovery, and it is pinned to `:app` by more than being a
+manifest entry point — it calls `WidgetUpdateScheduler.ensureScheduled`, and a `:core:*` module
+naming that is a `:core:*` → `:feature:*` edge `moduleBoundary` fails the build on. Splitting them
+is what makes the notification stack movable at all; `PrayerAlarmReceiver` names nothing in
+`:feature:widget`.
+
+Both receivers do their work inside `goAsync()`. Neither did before, and every branch of both
+starts work that outlives `onReceive` — so the process could be killed mid-reschedule, most likely
+right after a reboot, which is exactly when nothing else will do that work.
+
 ```text
 core/data/src/main/kotlin/             #  ← :core:data — the only module that sees both stores
 com.arshadshah.nimaz/
@@ -741,8 +753,8 @@ Report every failure through **`telemetry.failure(domain, type, throwable)`**, w
 channels — the stack trace to Crashlytics, the frequency to analytics — and ignores
 `CancellationException`, because a load cancelled by navigating away is not a failure. Calling
 `AppAnalytics.*` or `CrashReporter.*` directly from a ViewModel is a deviation; the objects remain
-only as the production binding and for callers with no injection point (`NimazApp`, `BootReceiver`,
-workers).
+only as the production binding and for callers with no injection point (`NimazApp`, the two
+broadcast receivers, workers).
 
 **Performance is on the seam too.** It was not, and the omission made the whole channel
 unreachable: `PerfMonitor` is a third object, this section forbids calling it, and `Telemetry`
@@ -1868,7 +1880,7 @@ eleven `:feature:*`, and `:app`:
 | **`:feature:quran`** | `nimaz.android.feature` | The reader, khatam and bookmarks, plus the whole Mushaf rendering stack. (`TajweedParser` came here in PR 19 and went on to `:core:ui` in PR 21, when `QuranSettingsScreen` took `TajweedLegendSheet` to a second feature module.) The largest feature. **`QuranDao` stays in `:core:database`** — four repositories use it. **`QuranAudioManager` stays in `:app`**, behind the `QuranPlayback` port, because `MainActivity` holds one too. |
 | **`:feature:prayer`** | `nimaz.android.feature` | When each prayer *is* and which way to face: prayer times, the monthly table, qibla and the night-worship window — the counterpart to `:feature:tracker`. The only module with a camera dependency (`ArQiblaView`). **The adhan players and the prayer notification machinery are *not* here**: nothing in the move set names them, and their consumers are the settings surface plus `:app` init, so sending them here would have created the `:feature:settings -> :feature:prayer` edge #571 forbids. **`PrayerTimeCard` and `PrayerSkyScene` went down to `:core:ui`**, being read by `HomeScreen`/`HomeHero` too. |
 | **`:feature:settings`** | `nimaz.android.feature` | The last feature module: 24 screens, the 1,400-line `SettingsViewModel`, location and sync. **Five screens arrive from other features' directories** — `DuaSettingsScreen`, `HadithSettingsScreen`, `SelectReciterScreen`, `SelectTranslationScreen`, `LocationScreen` — every one dispatching `SettingsEvent`. **`data/sync` did *not* come**: it imports 21 DAOs and 14 entities, so it went to `:core:data`. `PrayerNotificationScheduler` stayed in `:app`, pinned by one `AppR.drawable` line; the three members this module calls became the `PrayerAlarmScheduler` / `PrayerNotificationTester` ports. |
-| **`:app`** | `nimaz.android.application` | **53 files, 11,595 lines — 8% of the codebase.** What genuinely cannot leave: `MainActivity`, `NimazApp`, `NavGraph.kt`, six `core/di` modules, `core/init`, the notification stack (`PrayerNotificationScheduler`, `BootReceiver`, `PrayerRescheduler`, `NotificationContentHelper`, `PrayerAlarmTimes`), the two adhan Services, `QuranAudioService` and `LibraryRepositoryImpl` — most pinned by `com.arshadshah.nimaz.R` or by being a manifest entry point. Plus the **home surface**: `HomeScreen`, `HomeGraph`, `viewmodel/home`, and **21 components** (the eight `Home*` organisms — `HomeOccasionsSection`, which puts today's Islamic occasion on compact Home, is the newest, `EventsCarousel`, `TodayCarousel`, `TodayInfoCards`, `TodaysProgressCard`, `JumuahCard`, `WorshipEventCard`, `NimazCarousel`, and five molecules). Those components stayed for the inverse of the rule that moved so many others: Home is their *only* consumer, so there is no second module to share them with. |
+| **`:app`** | `nimaz.android.application` | **54 files, 11,711 lines — 8% of the codebase.** What genuinely cannot leave: `MainActivity`, `NimazApp`, `NavGraph.kt`, six `core/di` modules, `core/init`, the notification stack (`PrayerNotificationScheduler`, `PrayerAlarmReceiver`, `PrayerRescheduler`, `NotificationContentHelper`, `PrayerAlarmTimes`) and `BootReceiver`, which was split out of it, the two adhan Services, `QuranAudioService` and `LibraryRepositoryImpl` — most pinned by `com.arshadshah.nimaz.R` or by being a manifest entry point. Plus the **home surface**: `HomeScreen`, `HomeGraph`, `viewmodel/home`, and **21 components** (the eight `Home*` organisms — `HomeOccasionsSection`, which puts today's Islamic occasion on compact Home, is the newest, `EventsCarousel`, `TodayCarousel`, `TodayInfoCards`, `TodaysProgressCard`, `JumuahCard`, `WorshipEventCard`, `NimazCarousel`, and five molecules). Those components stayed for the inverse of the rule that moved so many others: Home is their *only* consumer, so there is no second module to share them with. |
 | **`:baselineprofile`** | `com.android.test` | Generates `app/src/main/baseline-prof.txt`. Nothing depends on it at runtime and no product code lives there. |
 
 Plus one **included build**, `build-logic`, which is not a module of the app — it produces the
