+++ docs/TESTING.md
@@ -1460,7 +1460,7 @@ production code changed and no `COVERAGE_EXCLUSIONS` entry added or widened.
 **Two things had to be built before a single test could run**, and both are reusable.
 
 `testing/TestEntryPointApplication.kt` is the smaller of them. `@AndroidEntryPoint` is not a
-no-op at run time: the Hilt Gradle plugin rewrites `BootReceiver.onReceive` and every
+no-op at run time: the Hilt Gradle plugin rewrites each `BroadcastReceiver.onReceive` and every
 `Service.onCreate` to inject first, and the generated base class then asks the **application** for
 a component — so an ordinary `android.app.Application` throws *"Hilt BroadcastReceiver must be
 attached to an @HiltAndroidApp Application"* before a line of the subject runs. That is why the
@@ -1488,7 +1488,7 @@ than in `build-logic`.
 | File | Pins |
 |---|---|
 | `core/util/PrayerNotificationSchedulerTest` (29) | Every alarm the app lives by, read back off `shadowOf(alarmManager).scheduledAlarms` rather than verified as a call. Nothing in the UI shows that an alarm was expected, so this class was at 1% while notifications were the app's headline feature. It pins that the midnight chain — the whole recurrence mechanism, since every other alarm is a one-shot — is armed for 00:01 tomorrow; that turning notifications off *cancels* rather than silently leaving yesterday's alarms armed; that no alarm is ever armed in the past, because Android delivers a past trigger immediately and the notification then re-posts on every reschedule for the rest of the day; and that a malformed stored khatam time falls back to 06:00 rather than throwing and taking every prayer alarm with it. Also the deliberate asymmetry `cancelAllPrayerNotifications` leaves behind: the daily summary is a recap rather than a prayer alert and survives it. |
-| `core/util/BootReceiverTest` (31), `BootReceiverEdgesTest` (16) | Where every prayer notification is actually produced. The alert style decides sound, and the tests assert on the *channel* because that is what Android reads: a silenced prayer lands on the muted channel at low priority, vibration off lands on the no-vibration sibling, and sunrise cannot be silenced at all because it is the end of Fajr's window rather than a prayer. The adhan path pins the rule that matters religiously — a missing variant falls back to the beep and **never** to the other variant, because playing the Fajr adhan at Dhuhr is wrong — and that Do Not Disturb gates the audio only, so a reader in a meeting is still told a prayer came in. Every handler is wrapped in `catch (e: Exception)`, which is right for a receiver with nowhere to propagate to and is also why each arm needs a test to exist at all: a handler that throws on its first line looks exactly like a quiet day. |
+| `core/util/PrayerAlarmReceiverTest` (30), `PrayerAlarmReceiverEdgesTest` (13), `BootReceiverTest` (8), `BootReceiverManifestTest` (4) | Where every prayer notification is actually produced. The alert style decides sound, and the tests assert on the *channel* because that is what Android reads: a silenced prayer lands on the muted channel at low priority, vibration off lands on the no-vibration sibling, and sunrise cannot be silenced at all because it is the end of Fajr's window rather than a prayer. The adhan path pins the rule that matters religiously — a missing variant falls back to the beep and **never** to the other variant, because playing the Fajr adhan at Dhuhr is wrong — and that Do Not Disturb gates the audio only, so a reader in a meeting is still told a prayer came in. Every handler is wrapped in `catch (e: Exception)`, which is right for a receiver with nowhere to propagate to and is also why each arm needs a test to exist at all: a handler that throws on its first line looks exactly like a quiet day. The recovery half is its own receiver and its own two files: `BootReceiverTest` covers the four actions that re-arm the alarms — including `MY_PACKAGE_REPLACED`, which is what carries an install across the rename of a `PendingIntent`'s target — and `BootReceiverManifestTest` checks the half of a receiver's contract that lives in XML. That second one exists because three branches of the old receiver read as supported, were covered by a unit test calling `onReceive` directly, and on a device never arrived once: there was no `<intent-filter>` behind them. |
 | `core/init/AppInitializerTest` (15) | The five startup tasks and the 5-second budget that lets the UI open regardless. A slow migration must not hold a reader at the splash screen when they opened the app to check Maghrib, and a task that throws must not stop the other four. Also that today's alarms are re-armed on every start — nothing else does it — and that an install which already has the audio does not re-download it on every cold start. |
 | `core/util/InAppUpdateManagerTest` (20) | The update banner's state machine, driven through a stubbed `AppUpdateManagerFactory`. The recoveries are the point: a failed check, a cancelled Play dialog and a flow Play refuses to launch each have to leave the banner interactive, because a banner stuck on a spinner is a dead control with no other route to the update. Play's callbacks post to the main looper, so every assertion idles it first — without that the whole class passes reading `Checking`. |
 | `data/audio/AdhanPlaybackServiceTest` (14), `AdhanDownloadServiceTest` (13), `QuranAudioServiceTest` (15) | The three foreground services, each built with `Robolectric.buildService`. What they produce is a notification, and the endings are what a reader sees when something half-works: a download that fetched one variant is reported as **partial** rather than done, because "downloaded" with no Fajr file means silence at the one prayer nobody is awake to notice. The Quran service's notification has to follow the audio state — Pause while playing, ongoing only while playing, download progress instead of the reciter while preparing — and take itself down when audio ends rather than sitting in the shade. |
